sprint_7-solution-in-file-manager v2
